backend/requirements.txt
fastapi==0.115.0
email-validator==2.2.0
uvicorn==0.30.6
sqlalchemy==2.0.36
alembic==1.13.2
celery==5.4.0
redis==5.0.8
psycopg2-binary==2.9.9
asyncpg==0.30.0
python-dotenv==1.0.1
pydantic-settings==2.5.2
google-generativeai==0.8.4
dnspython==2.6.1
requests==2.32.3
# Requires the `nmap` system binary on PATH (Celery worker host). See README / run.sh ensure_nmap.
python-nmap==0.7.1
resend==2.4.0
supabase==2.8.1
